requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
asgiref==3.6.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
discord.py==1.7.3
dj-database-url==2.0.0
Django==4.2
django-heroku==0.3.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gunicorn==20.1.0
h5py==3.6.0
idna==2.10
importlib-metadata==4.8.2
inflection==0.5.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
keyboard==0.13.5
kiwisolver==1.3.1
libclang==12.0.0
Markdown==3.3.6
more-itertools==8.8.0
MouseInfo==0.1.3
multidict==5.1.0
oauthlib==3.1.1
opt-einsum==3.3.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.19.1
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pygame==2.0.1
PyGetWindow==0.0.9
pyglet==1.5.15
Pygments==2.9.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2021.1
pywin32==301
Quandl==3.6.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
threadpoolctl==2.1.0
traitlets==5.0.5
typing-extensions==3.10.0.0
tzdata==2023.3
urllib3==1.26.5
vboxapi==1.0
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==6.4.0
wrapt==1.13.3
yarl==1.6.3
zipp==3.6.0
